requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
faiss-cpu==1.9.0.post1
filelock==3.20.0
Flask==2.3.2
Flask-Cors==4.0.0
fsspec==2025.10.0
huggingface-hub>=0.20.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.2
numpy>=1.26.0
packaging==25.0
pandas>=2.1.0
pillow==12.0.0
psutil==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn>=1.3.2
imbalanced-learn>=0.11.0
scipy>=1.11.4
# sentence-transformers>=3.0.0  <-- Removed
# torch>=2.1.0                  <-- Removed
# torchvision>=0.16.0           <-- Removed
onnxruntime>=1.16.0
transformers>=4.40.0
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers>=0.19.0
tqdm==4.67.1
typing_extensions>=4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.6
